+++ R/app_version.R
@@ -1,1 +1,1 @@
-app_version <- function()'25.10.3'
+app_version <- function()'25.10.4'