+++ R/hosted_version.R
@@ -1,1 +1,1 @@
-hosted_version <- function()'v25.10.3-251021'
+hosted_version <- function()'v25.10.4-251027'